automatically abort when a deadlock is detected
This changes the architecture of the module. Previously, the module was
only exposing an AMI command "CheckFreeze", which was telling if
asterisk was deadlocked or not. In XiVO, it was monit which was
responsible to call this AMI command regularly and abort the process if a
deadlock was detected.

Now, when the module is loaded, it starts a thread that check at regular
interval (60 seconds) if asterisk is deadlocked, and if it is, it
terminates the process with a ABRT signal.

At the same time, I've also changed a bit how mutexes are tested by the
module: it now only needs a reference of the ast_mutex to test, and it
does the deadlock test using pthread_mutex_timedlock, which is safer
(and a bit more efficient) than doing the check with multiple
pthread_mutex_trylock calls. This would also make it easier to test more
mutexes than just the one associated to the global channels container.