remove channels lock check
Why: with the introduction of CHANNELSTORAGE_API in 22.5.0, locks
are exposed via this API and it requires a lot of work to do the
same checks as we did before. Given that we will switch to the
new channel backend soon, it does not make sense to migrate the
freeze checks when we will not need them with the new backend.